% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_balance_energy.R
\name{fdk_balance_energy}
\alias{fdk_balance_energy}
\title{Energy balance closure}
\usage{
fdk_balance_energy(qle, qle_qc, qh, qh_qc, rnet, qg, qg_qc, time, tstepsize)
}
\arguments{
\item{qle}{latent heat}

\item{qle_qc}{quality control}

\item{qh}{sensible heat}

\item{qh_qc}{quality control}

\item{rnet}{respiration?}

\item{qg}{not sure check}

\item{qg_qc}{quality control}

\item{time}{time field?}

\item{tstepsize}{time step}
}
\value{
sensible and latent heat values with the energy balance closed
}
\description{
Close the energy balance of flux measurements.
Takes a netcdf input file as generated by dfk_process_lsm() which
calls fdk_correct_fluxes()
}
