% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_correct_rs.R
\name{fdk_correct_rs}
\alias{fdk_correct_rs}
\title{Rename remote sensing variables}
\usage{
fdk_correct_rs(infile)
}
\arguments{
\item{infile}{netcdf file (path) holding the ERA and remote
sensing data}
}
\value{
renames netcdf MODIS data variables
}
\description{
Rename remote sensing variables in the ancillary netcdf data
as created by FluxnetLSM.
}
