% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_data_coverage.R
\name{fdk_data_coverge}
\alias{fdk_data_coverge}
\title{Calculates data coverage}
\usage{
fdk_data_coverge(
  input_path = "/data/scratch/beta-v3/lsm/",
  output_path,
  internal = TRUE
)
}
\arguments{
\item{input_path}{path with half-hourly gap-filled plumber formatted data}

\item{output_path}{where to store the converted data if converted to csv}

\item{internal}{returns an internal data frame or a csv written to file
named 'data_coverage.csv' (either TRUE or FALSE)}
}
\value{
data frame or csv written to disk with data coverage statistics
 per site
}
\description{
Calculates data coverage for the half-hourly (HH) data.
}
