% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_detect_outliers.R
\name{fdk_detect_outliers}
\alias{fdk_detect_outliers}
\title{Detect outliers in a time series}
\usage{
fdk_detect_outliers(df, iterations = 20, sigma = 2, plot = FALSE)
}
\arguments{
\item{df}{dataframe with dates (date) and values (value) to smooth}

\item{iterations}{number of itterations in order to detect outliers ()}

\item{sigma}{number of deviations to exclude outliers at}

\item{plot}{visualize the process, mostly for debugging
(\code{TRUE} / \code{FALSE} = default)}
}
\description{
The function fills in the existing column to hold outlier flags,
and either overwrites the original file or outputs a data structure.
}
