% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_download_cloud_cover.R
\name{fdk_download_cloud_cover}
\alias{fdk_download_cloud_cover}
\title{Download ERA5 cloud cover data}
\usage{
fdk_download_cloud_cover(df, user, path = "data-raw/cloud_cover/")
}
\arguments{
\item{df}{data frame with site info}

\item{user}{ECMWF CDS user name}

\item{path}{path where to store the ERA5 data}
}
\value{
raw cloud cover data saved to a desired path
}
\description{
Download ERA5 cloud cover data from the ECMWF Copernicus Data Service
}
