% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_downsample_fluxnet_phydro.R
\name{fdk_downsample_fluxnet_phydro}
\alias{fdk_downsample_fluxnet_phydro}
\title{Down samples half-hourly FLUXNET data to daily values}
\usage{
fdk_downsample_fluxnet_phydro(
  df,
  site,
  out_path,
  overwrite = FALSE,
  save_plots = TRUE,
  method = c("24hr", "daytime", "3hrmax")[1],
  midday_criterion = c("quantile", "time")[2]
)
}
\arguments{
\item{df}{a half-hourly FLUXNET data frame}

\item{site}{sitename of a site to process}

\item{out_path}{the path where to store the converted data}

\item{overwrite}{overwrite existing output file}

\item{save_plots}{Whether plots of downsampled variables before and after gapfilling should be saved to files (in the output_folder OR to default device)}

\item{method}{What method should be used for averaging (should be one of "24hr", "daytime", "3hrmax")}

\item{midday_criterion}{How should midday time points be identified? Should be one of "quantile", "time"}
}
\value{
data frame with daily (DD) down sampled values or file in the
 FLUXNET format
}
\description{
Normally the flux data is generated using a dedicated flux processing
workflow. Therefore the data are not 100% identical but correspondence
should have an R2 > 0.95, with only auto-correlated errors in the NT product.
}
\details{
This function and the data generated should be used with caution, knowing
the limitations that this routine implies with respect to all variables
included.
}
