% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_format_drivers_phydro.R
\name{fdk_format_drivers_phydro}
\alias{fdk_format_drivers_phydro}
\title{Format p-model drivers}
\usage{
fdk_format_drivers_phydro(
  site_info,
  params_siml = dplyr::tibble(spinup = TRUE, spinupyears = 10, recycle = 1, outdt = 1,
    ltre = FALSE, ltne = FALSE, ltrd = FALSE, ltnd = FALSE, lgr3 = TRUE, lgn3 = FALSE,
    lgr4 = FALSE, use_phydro = TRUE, use_pml = FALSE),
  path,
  verbose = TRUE,
  fdk_inputs_dir = "/data_2/FDK_inputs/"
)
}
\arguments{
\item{site_info}{data frame using minimum information required
being five columns: sitename, lon, lat, elv, whc}

\item{params_siml}{simulation parameters (preset)}

\item{path}{path with daily FLUXNET data}

\item{verbose}{provide verbose output (default = FALSE)}

\item{product}{which flux product to use}
}
\value{
returns an rsofun compatible driver file for the provided
 sites
}
\description{
Takes site information for a single
or multiple sites and grabs all data
required for a p-model run in rsofun.
}
\details{
Parameter settings are provided as
arguments, but could be altered after
the fact if desired.

NOTE: Processing is selective and will
create different files on non-GECO workstations.
}
