% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_get_sequence.R
\name{fdk_get_sequence}
\alias{fdk_get_sequence}
\title{Get sequence of good-quality data}
\usage{
fdk_get_sequence(
  df,
  site = "sitename",
  qc_threshold = 0.5,
  leng_threshold = 25,
  do_plot = FALSE,
  out_path = "data/tmp"
)
}
\arguments{
\item{df}{data frame with daily FLUXNET-based data}

\item{site}{a FLUXNET site name to process}

\item{qc_threshold}{Threshold of the the QC information (fraction of good-
quality half-hourly data used for daily aggregates).}

\item{leng_threshold}{Threshold of interruptions of bad-quality data to be
ignored. In days.}

\item{do_plot}{boolean, whether to create plot, defaults to FALSE}

\item{out_path}{the path where to store the plot}
}
\value{
A data frame containing information about the start date and end date
of the longest sequence.
}
\description{
Determines the longest consecutive sequence of good-quality data, separately
for GPP and LE. Short "interruptions" of bad-quality data are ignored as long
as they don't exceed the given length threshold
}
