% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fdk_impute_knn}
\alias{fdk_impute_knn}
\title{Impute with KNN}
\usage{
fdk_impute_knn(df, target, pred1, pred2 = NULL, pred3 = NULL, k)
}
\arguments{
\item{df}{a data frame containing columns corresponding to the target and
predictor variable names}

\item{target}{a string specifying the target variable (column) name}

\item{pred1}{a string specifying the first predictor variable (column) name}

\item{pred2}{a string specifying the second predictor variable (column) name}

\item{pred3}{a string specifying the third predictor variable (column) name}

\item{k}{an integer specifying the number of neighbours to consider}
}
\value{
a gap filled data frame
}
\description{
Impute missing values with a K-nearest neighbour
}
