% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_download_modis.R
\name{fdk_match_modis}
\alias{fdk_match_modis}
\title{match MODIS LAI/FPAR to ERA data}
\usage{
fdk_match_modis(df, path, nc_file)
}
\arguments{
\item{df}{data frame with site info}

\item{path}{path where to store the MODIS data}

\item{nc_file}{netcdf file containing the LSM processed flux data}
}
\value{
smoothed time series of LAI/FPAR integrated in the meteorological
(ERA) data
}
\description{
Downloads and smooths MODIS LAI/FPAR values
for merging into the final LSM model data.
Smoothing interpolates values to an half-hourly
time step.
}
