% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_optimal_span.R
\name{fdk_optimal_span}
\alias{fdk_optimal_span}
\title{Calculates the optimal span for a loess spline}
\usage{
fdk_optimal_span(
  y,
  x = NULL,
  weights = NULL,
  step = 0.01,
  label = NULL,
  plot = FALSE
)
}
\arguments{
\item{y}{a vector with measurement values to smooth}

\item{x}{a vector with dates / time steps}

\item{weights}{optional values to weigh the loess fit with}

\item{step}{span increment size}

\item{label}{title to be used when plotting function output}

\item{plot}{plot visual output of the optimization routine}
}
\value{
Returns an optimal span to smooth a provided vector using
the `loess()` smoother.
}
\description{
The optimal span is calculated based upon the Bayesian information
criterion (BIC). This is a copy of the routine used in the
{phenocamr} package.
}
