% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_plot.R
\name{fdk_plot}
\alias{fdk_plot}
\title{Plot FluxDataKit output}
\usage{
fdk_plot(site, path = "data/tmp", out_path = "data/tmp")
}
\arguments{
\item{site}{a sitename}

\item{path}{the path where LSM compatible data are}

\item{out_path}{where to store the images
stored}
}
\value{
a plot with key variables
}
\description{
Plot FluxDataKit output
}
