% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_process_cloud_cover.R
\name{fdk_process_cloud_cover}
\alias{fdk_process_cloud_cover}
\title{Process ERA5 cloud cover data}
\usage{
fdk_process_cloud_cover(path = "data-raw/cloud_cover/", site = "CA-NS4")
}
\arguments{
\item{path}{path with the ERA5 cloud cover data, by site}

\item{site}{site name to process}

\item{start_date}{start date of a data series}

\item{end_date}{end date of a data series}
}
\value{
daily mean cloud cover value (0-1), for a given date range
}
\description{
Process cloud cover data to return daily cloud cover values
for a given date range and (flux) site.
}
