% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_process_lsm.R
\name{fdk_process_lsm}
\alias{fdk_process_lsm}
\title{Generate LSM data}
\usage{
fdk_process_lsm(
  df,
  out_path,
  modis_path,
  format = "lsm",
  save_tmp_files = TRUE,
  overwrite = TRUE,
  site_csv_file = system.file("extdata", "Site_metadata.csv", package = "FluxnetLSM")
)
}
\arguments{
\item{df}{dataframe with sites to process}

\item{out_path}{output directory for processed data}

\item{modis_path}{where to store downloaded MODIS data}

\item{format}{the format of the output (fluxnet = FLUXNET formatting)}

\item{save_tmp_files}{retain temporary files (TRUE or FALSE)}

\item{site_csv_file}{CSV file with site meta-data which is parsed to populate
ancillary data provided with the flux data. By default the hard coded
packaged data are used, but alternatively an external file can be
provided. This limits the need to recompile the package when processing
files which were not originally selected.}
}
\value{
LSM compatible netcdf files in the output directory, with
 intermediary files saved upon request
}
\description{
Convert FLUXNET data to land surface model compatible netcdf files, using
the FluxnetLSM package and PLUMBER2 based workflows.
}
