% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_release.R
\name{fdk_release}
\alias{fdk_release}
\title{Creates a new FluxDataKit product release}
\usage{
fdk_release(
  df,
  input_path,
  output_path,
  overwrite_lsm = TRUE,
  overwrite_fluxnet = TRUE
)
}
\arguments{
\item{df}{data frame with site meta-data information}

\item{input_path}{path with input data}

\item{output_path}{path where to store the output data}

\item{overwrite_lsm}{overwrite the LSM-formatted NetCDF files if the output
path exists}

\item{overwrite_fluxnet}{overwrite the FLUXNET-formatted CSV files if the
output path exists}
}
\value{
the FluxDataKit data release as a set of netCDF, CSV data,
 compressed RDS rsofun drivers and meta-data
}
\description{
Creates a new product release when data is organized
in the fixed data structure as specified. This function
calls individual functions which are more flexible but
for consistency this wrapper restricts things further.
}
