% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_setup_sofun.R
\name{fdk_setup_sofun}
\alias{fdk_setup_sofun}
\title{Complements the setup settings}
\usage{
fdk_setup_sofun(site_info, params_siml)
}
\arguments{
\item{site_info}{A character string specifying the path to the site meta
info file, or a dataframe containing the site meta info.}

\item{params_siml}{A named list containing the simulation parameters
for SOFUN.}
}
\value{
A dataframe (tibble) containing the site meta info,
complemented by column 'params_siml' which is a nested list
of complemented simulation parameters.
}
\description{
Complements the settings based on the site meta info CSV file or data frame.
}
\examples{
\dontrun{
 setup <- prepare_setup_sofun(
   site_info = site_info,
   params_siml = params_siml
  )
}
}
