% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fdk_site_info}
\alias{fdk_site_info}
\title{FDK site info file}
\format{
DataFrame
}
\usage{
fdk_site_info
}
\description{
List of sites to processes as well
as their meta-data
}
\keyword{datasets}
