% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_smooth_ts.R
\name{fdk_smooth_ts}
\alias{fdk_smooth_ts}
\title{Smooth MODIS time series}
\usage{
fdk_smooth_ts(df, variable, nc_file)
}
\arguments{
\item{df}{vector of dates}

\item{variable}{what MODIS variable is processed (LAI or FPAR)}

\item{nc_file}{FluxnetLSM netcdf ERA Met file to amend, scaling
the data to the respective time step required}
}
\value{
a data frame with smoothed gapfilled time series
}
\description{
Smooth and gapfill modis time series, and insert
results into FluxnetLSM netcdf files for data fusion.
}
