% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_correct_fluxes.R
\name{flux_corrections}
\alias{flux_corrections}
\title{Flux corrections}
\usage{
flux_corrections(
  infile_flux,
  qc_info,
  outfile_flux,
  outdir,
  new_qc,
  qle_name = "Qle",
  qh_name = "Qh",
  rnet_name = "Rnet",
  qg_name = "Qg",
  qle_cor_name = "Qle_cor",
  qh_cor_name = "Qh_cor"
)
}
\arguments{
\item{qh_cor_name}{}
}
\value{

}
\description{
Additional flux corrections
}
