% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{interpolate2daily_CO2_F_MDS}
\alias{interpolate2daily_CO2_F_MDS}
\title{Interpolate missing CO2 data}
\usage{
interpolate2daily_CO2_F_MDS(df)
}
\arguments{
\item{df}{a data frame containing missing values for CO2_F_MDS}
}
\value{
a gap filled data frame
}
\description{
Interpolate missing CO2 data
}
