% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{interpolate2daily_fpar}
\alias{interpolate2daily_fpar}
\title{Interpolate missing FPAR and LAI data}
\usage{
interpolate2daily_fpar(df)
}
\arguments{
\item{df}{a data frame containing missing values for FPAR and LAI}
}
\value{
a gap filled data frame
}
\description{
Interpolate missing FPAR and LAI data
}
