% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{interpolate2daily_PA_F}
\alias{interpolate2daily_PA_F}
\title{Interpolate missing atmospheric pressure data}
\usage{
interpolate2daily_PA_F(df)
}
\arguments{
\item{df}{a data frame containing missing values for CO2_F_MDS}
}
\value{
a gap filled data frame
}
\description{
Interpolate missing atmospheric pressure data
}
