% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{interpolate2daily_WS_F}
\alias{interpolate2daily_WS_F}
\title{Interpolate missing wind speed data}
\usage{
interpolate2daily_WS_F(df)
}
\arguments{
\item{df}{a data frame containing missing values for FPAR and LAI}
}
\value{
a gap filled data frame
}
\description{
Interpolate missing wind speed data
}
