% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_exceptions.R
\name{site_exceptions}
\alias{site_exceptions}
\title{Address ad-hoc data exceptions in conversions}
\usage{
site_exceptions(site_code, var_data, att_data, qc_val)
}
\arguments{
\item{site_code}{FLUXNET site code abbreviation}

\item{var_data}{FLUXNET data frame}

\item{att_data}{FLUXNET meta-data}

\item{qc_val}{FLUXNET quality control parameters}
}
\value{
corrected FLUXNET data for select sites
}
\description{
Despite standardization efforts it seems that some flux data has
idiosyncratic issues with respect to units etc. For the few sites
where such issues exist dedicated correction routines are provided.
}
\details{
This is an almost verbatim copy of the routine as used to generate
the PLUMBER-2 data. For transparency reasons this part was split out into
a function to provide easier debugging and processing options.
}
