% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{SynthesizeLWdown}
\alias{SynthesizeLWdown}
\title{Synthesises downward longwave radiation based on Tair and rel humidity}
\usage{
SynthesizeLWdown(TairK, RH, technique = "Abramowitz_2012")
}
\description{
Synthesises downward longwave radiation based on Tair and rel humidity
}
