% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{VPD2RelHum}
\alias{VPD2RelHum}
\title{Converts VPD (hPa) to relative humidity (percentage)}
\usage{
VPD2RelHum(VPD, airtemp, vpd_units, tair_units)
}
\description{
Converts VPD (hPa) to relative humidity (percentage)
}
