% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_balance_energy.R
\name{fdk_balance_energy}
\alias{fdk_balance_energy}
\title{Energy balance closure}
\usage{
fdk_balance_energy(qle, qle_qc, qh, qh_qc, rnet, qg, qg_qc, time, tstepsize)
}
\arguments{
\item{qle}{latent heat}

\item{qle_qc}{latent heat quality control}

\item{qh}{sensible heat}

\item{qh_qc}{sensible heat quality control}

\item{rnet}{respiration rate}

\item{qg}{not sure check}

\item{qg_qc}{quality control parameter}

\item{time}{a date/time field (vector)}

\item{tstepsize}{time step length}
}
\value{
sensible and latent heat values with the energy balance closed
}
\description{
Close the energy balance of flux measurements.
Takes a netcdf input file as generated by dfk_process_lsm() which
calls fdk_correct_fluxes()
}
\details{
This is an almost verbatim copy of the routine as used to generate
the PLUMBER-2 data. For transparency reasons this part was split out into
a function to provide easier debugging and processing options.
}
