% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_balance_energy.R
\name{fdk_balance_energy}
\alias{fdk_balance_energy}
\title{Energy balance closure}
\usage{
fdk_balance_energy(qle, qle_qc, qh, qh_qc, rnet, qg, qg_qc, time, tstepsize)
}
\arguments{
\item{tstepsize}{}
}
\description{
Close the energy balance of flux measurements.
Takes a netcdf input file as generated by dfk_process_lsm()
}
