% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_convert_lsm.R
\name{fdk_convert_lsm}
\alias{fdk_convert_lsm}
\title{Converts LSM NetCDF data to FLUXNET}
\usage{
fdk_convert_lsm(
  site,
  path,
  fluxnet_format = FALSE,
  meta_data = FALSE,
  out_path,
  overwrite = FALSE
)
}
\arguments{
\item{site}{fluxnet site name}

\item{path}{path with plumber2 data (both flux and meteo data files)}

\item{fluxnet_format}{convert to fluxnet formatting (TRUE/FALSE)}

\item{meta_data}{return meta-data TRUE/FALSE}

\item{out_path}{where to store the converted data if converted to
fluxnet formatting and returns both half-hourly and daily data.}

\item{overwrite}{overwriting existing files (TRUE/FALSE). Defaults to FALSE.}
}
\value{
data frame with merged meteo and flux data
}
\description{
Reads LSM data (as a netcdf) in a given directory
by (fluxnet) site name, optionally only returns the meta-data of the
site data.
}
\details{
Default parameters return a CSV file which is compatible with the current
OneFlux/FLUXNET naming conventions. Returned data will retain its
half-hourly (HH) time step. For aggregation to a daily time step consider
the `fdk_downsample_fluxnet()` function.
}
