% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_correct_era.R
\name{fdk_correct_era}
\alias{fdk_correct_era}
\title{Corrects ERA data}
\usage{
fdk_correct_era(infile_met, new_qc)
}
\arguments{
\item{infile_met}{input netcdf filename}

\item{new_qc}{quality control parameters}
}
\value{
corrects an existing ERA-Interim FluxnetLSM file in place
}
\description{
Corrects ERA data of a FluxnetLSM file (changes the file)
locally, if a path is provided a new file is created.
}
