% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_download_modis.R
\name{fdk_download_modis}
\alias{fdk_download_modis}
\title{Download MODIS LAI/FPAR}
\usage{
fdk_download_modis(df, path)
}
\arguments{
\item{path}{}
}
\description{
Download MODIS LAI/FPAR data and save data to file
or return internally
}
