% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_downsample_fluxnet.R
\name{fdk_downsample_fluxnet}
\alias{fdk_downsample_fluxnet}
\title{Down samples half-hourly FLUXNET data to daily values}
\usage{
fdk_downsample_fluxnet(df, site, out_path, overwrite = FALSE)
}
\arguments{
\item{df}{a half-hourly FLUXNET data frame}

\item{site}{sitename of a site to process}

\item{out_path}{the path where to store the converted data}

\item{overwrite}{overwrite existing output file}
}
\value{
data frame with daily (DD) down sampled values or file in the
 FLUXNET format
}
\description{
Down sampling of GPP values is done at-hoc using a simple
mean.
}
\details{
Normally the data is generated using a dedicated flux processing
workflow. Therefore the data are not 100% identical but correspondence
should have an R2 > 0.95, with only structured errors in the NT product.

This function and the data generated should be used with caution, knowing
the limitations that this routine implies with respect to all variables
included.
}
