% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_downsample_fluxnet.R
\name{fdk_downsample_fluxnet}
\alias{fdk_downsample_fluxnet}
\title{Down sample half-hourly data to daily values}
\usage{
fdk_downsample_fluxnet(df, site, out_path)
}
\arguments{
\item{df}{FLUXNET based HH data frame}

\item{out_path}{where to store the converted data if converted to
fluxnet formatting}
}
\value{
data frame with daily (DD) down sampled values
}
\description{
Currently CO2 fluxes can not be down sampled using the
FLUXNET routine due to missing methodology.
}
