% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_era5_request.R
\name{fdk_era5_request}
\alias{fdk_era5_request}
\title{Download request for drivers from ECMWF}
\usage{
fdk_era5_request(lat, lon, filename, start_date, end_date)
}
\arguments{
\item{lat}{latitude}

\item{lon}{longitude}

\item{filename}{output filename (CSV)}

\item{start_date}{start date as YYYY-MM-DD}

\item{end_date}{end date as YYYY-MM-DD}
}
\value{
A data frame with queried data including the date, position and
requested variable
}
\description{
Batch request for ERA5 driver data from the ECMWF API (CDS Toolbox)
}
