% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_correct_fluxes.R
\name{fdk_flux_corrections}
\alias{fdk_flux_corrections}
\title{Flux corrections routine}
\usage{
fdk_flux_corrections(
  infile,
  qle_name = "Qle",
  qh_name = "Qh",
  rnet_name = "Rnet",
  qg_name = "Qg",
  qle_cor_name = "Qle_cor",
  qh_cor_name = "Qh_cor"
)
}
\arguments{
\item{infile}{input netcdf file}

\item{qle_name}{latent heat}

\item{qh_name}{sensible heat}

\item{rnet_name}{respiration flux}

\item{qg_name}{quality control}

\item{qle_cor_name}{quality control}

\item{qh_cor_name}{quality control}
}
\value{
energy balance corrected netcdf file
}
\description{
Wrapper for the energy balance correction and subsetting of
reported (valid) years.
}
\details{
This is an almost verbatim copy of the routine as used to generate
the PLUMBER-2 data. For transparency reasons this part was split out into
a function to provide easier debugging and processing options.
}
