% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_plot.R
\name{fdk_plot}
\alias{fdk_plot}
\title{Plot FluxDataKit output}
\usage{
fdk_plot(file, out_path = "data/tmp", overwrite = FALSE)
}
\arguments{
\item{file}{a fdk file}

\item{out_path}{where to store the images}

\item{overwrite}{overwrite existing files (TRUE or FALSE)
stored}
}
\value{
a plot with key variables
}
\description{
Plot FluxDataKit output
}
