% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_plot.R
\name{fdk_plot}
\alias{fdk_plot}
\title{Plot FluxDataKit output}
\usage{
fdk_plot(
  df,
  site = "sitename",
  out_path = "data/tmp",
  overwrite = FALSE,
  daily = FALSE
)
}
\arguments{
\item{df}{data frame with FLUXNET based data}

\item{site}{a FLUXNET site name to process}

\item{out_path}{the path where to store the plot}

\item{overwrite}{overwrite existing files (TRUE or FALSE)
stored}

\item{daily}{whether daily or hourly data is to be plotted, defaults to FALSE.}
}
\value{
A plot with FLUXNET variables for rapid visual data inspection
}
\description{
Simple routine to generate an overview plot of all
flux and climate data. The function takes a data frame
of FLUXNET formatted columns as input (this can be either
original source material or gap filled data as generated
by FluxDataKit).
}
\details{
Plots are not returned (shown) in the current interface
but only written to file to allow for headless batch
processing (with limited overhead).
}
