% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_process_lsm.R
\name{fdk_process_lsm}
\alias{fdk_process_lsm}
\title{Generate LSM data}
\usage{
fdk_process_lsm(
  df,
  out_path,
  format = "fluxnet",
  save_tmp_files = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{df}{dataframe with sites to process}

\item{out_path}{output directory}

\item{format}{the format of the output (fluxnet = FLUXNET formatting)}

\item{save_tmp_files}{retain temporary files (TRUE or FALSE)}
}
\description{
Convert FLUXNET data to land surface model compatible netcdf files, using
the FluxnetLSM package and PLUMBER2 based workflows.
}
