% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_process_lsm.R
\name{fdk_process_lsm}
\alias{fdk_process_lsm}
\title{Generate LSM data}
\usage{
fdk_process_lsm(
  df,
  out_path,
  modis_path,
  save_tmp_files = FALSE,
  overwrite = TRUE
)
}
\arguments{
\item{df}{dataframe with sites to process}

\item{out_path}{output directory for processed data}

\item{modis_path}{where to store downloaded MODIS data}

\item{save_tmp_files}{retain temporary files (TRUE or FALSE)}
}
\value{
LSM compatible netcdf files in the output directory, with
 intermediary files saved upon request
}
\description{
Convert FLUXNET data to land surface model compatible netcdf files, using
the FluxnetLSM package and PLUMBER2 based workflows.
}
