% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_read_plumber.R
\name{fdk_read_plumber}
\alias{fdk_read_plumber}
\title{Reads (Plumber) NETCDF files}
\usage{
fdk_read_plumber(file, meta_data = FALSE)
}
\arguments{
\item{file}{plumber netcdf file}
}
\value{
dataframe with requested values (meteo, flux or meta-data)
}
\description{
Returns either the raw flux or meteo
data or the site meta-data based upon
a provided path to a plumber Met or Flux
file.
}
