% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fdk_site_info}
\alias{fdk_site_info}
\title{FDK site info file}
\format{
DataFrame
\describe{
  \item{sitename}{Site name}
  \item{lon}{Longitude of site}
  \item{lat}{Latitude of site}
  \item{elv}{elevation}
  \item{year_start}{start year of data availability}
  \item{year_end}{end year of data availability}
  \item{koeppen_code}{Koeppen Geiger code}
  \item{igbp_land_use}{IGBP land cover type}
  \item{whc}{Total root zone water holding capacity}
  \item{product}{Data taken from published product}
}
}
\usage{
fdk_site_info
}
\description{
List of sites to processes as well
as their meta-data
}
\keyword{datasets}
