% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdk_smooth_ts.R
\name{fdk_smooth_ts}
\alias{fdk_smooth_ts}
\title{Smooth MODIS time series}
\usage{
fdk_smooth_ts(dates, values, variable, start_year, end_year, nc_file)
}
\arguments{
\item{dates}{vector of dates}

\item{values}{vector of values}

\item{variable}{what MODIS variable is processed (LAI or FPAR)}

\item{start_year}{a start year of the desired series}

\item{end_year}{an end year of the desired series}

\item{nc_file}{FluxnetLSM netcdf file to amend, scaling
the data to the respective time step required}
}
\value{
a data frame with smoothed gapfilled time series
}
\description{
Smooth and gapfill modis time series, and insert
results into FluxnetLSM netcdf files for data fusion.
}
