% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fill_netrad}
\alias{fill_netrad}
\title{Gap fill net radiation measurements}
\usage{
fill_netrad(df)
}
\arguments{
\item{df}{a data frame containing a netrad data column, and temperature
and light variables}
}
\value{
a gap filled data frame
}
\description{
Gap fill net radiation measurements using a
KNN approach, using temperature and ppfd
}
